#213 flag-it button not appearing correctly in alpha003 build
The fix, yesterday, to the spelling of "appearance" caused the Gallery's
theme's appearance for the flag-it button to finally be used correctly. It had
garbage in it for testing purposes.

This commit fixes that theme entry. It also removes the superfluous class
aiagallery.widget.Button which was created for this button, and instead
assigns the appearance name directly to the qx.ui.form.Button instance.

Derrell